requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
passlib
python-jose[cryptography]
passlib[bcrypt]
python-multipart